+++ example/custom/lib/assets/r.dart
@@ -1,0 +1,5 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+abstract final class R {
+  static const String assetsResponse = 'assets/response.json';
+}