+++ example/custom/lib/assets/random_images.dart
@@ -1,0 +1,5 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+abstract final class RandomImages {
+  static const String shieldHero = 'assets/images/shield_hero.jpg';
+}