+++ example/default/lib/assets.dart
@@ -1,0 +1,7 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+abstract final class Assets {
+  static const String assetsMainImageJpg = 'assets/main_image.jpg';
+
+  static const String assetsResponseJson = 'assets/response.json';
+}