+++ example/default/lib/images.dart
@@ -1,0 +1,5 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+abstract final class Images {
+  static const String shieldHero = 'assets/images/shield_hero.jpg';
+}