+++ lib/src/data/fatal_error.dart
@@ -1,0 +1,8 @@
+class StateException implements Exception {
+  final String message;
+
+  StateException(this.message);
+
+  @override
+  String toString() => message;
+}